Intro to Python Final Sol (for coursera review)
The solution to the final project of the intro to python programming by the University of Pennsylvania. The course project is necessarily complex here for review purposes. But here I have figured out the solution. I hope it's going to help you if you search for the answer. 

Just don't copy and paste completely. Thanks.